DO NOT MERGE Backport to latest calendar sync changes.
The CalendarImporterTest in this CL contains an API fix. In ag/10400533
this file was reverted as it used APIs that aren't available in
qt-qpr1-dev. The backported file does now contain a fix for this and
uses only APIs available.

In addition to the API fix of the unit test, this contains two commit
that were missed when the companiondevicesupport package got backported:
ag/10296958
ag/10296960

Bug: 150166597
Test: Unit tests for CompanionDeviceSupport pass.
Change-Id: I8b7258ab397a9c3391e5909e2b0552b7ec2423a2