Suppress notifications while projection is active
If a projection app, like AAP or Carplay is in the active foreground, we
should just post message notifications that originated from the phone
(and hence will already be shown in the projection app) silently in
the notification center without posting a HUN.

Bug: 138747359
Test: verified on hawk. CDS + AAP was not working together well, so I
mocked the changing of the Projection Status, verified that convos
posted with HUNs when Projection is not active, and the same + new
convos posted silently when Projection is in background. Also tested
the other way around.

Change-Id: I21368f19788ffc9d7c6a7c4ef6c892c2294bf5f4