Untie the AssociationActivity lifecycle from Settings
Currently, if AssociationActivity is entered from Settings, when Settings is brought to the foreground again, the AssociationActivity will be destroyed. As Settings is the only entry to AssociationActivity for users, we need to untie its lifecycle from Settings, so users can continue association if they leave the activity and get back during association.

Bug: 150945672
Test: Enter AssociationActivity from Settings, start pairing flow until the pairing code appears, leave the activity, and enter the activity again from Settings. The activity is still on pairing code screen, and association can be completed by tapping confirm button.
Change-Id: I824a31f5545dfa80b2b44fd95c01e2f28083e302