Prevents crash due to timing of when each process is killed
According to our failure analysis, crashes occurred on this line at the rate of 1 per 100, which significantly affected the behavior of Companion Device.
This crash was confirmed to occur when user switch is performed via key fob. So, we added error handling to prevent this crash when a secondary user process is killed by the system when an AIDL method has not finished running.

PIT: 301902990
Bug: 301902990
Change-Id: I7e73f7d9398e02236f6f9b77b33e448737efb2c2